Use font-lock flushing only interactively
Apparently the phenomenon of font-lock flushing being a necessity to
enable changes to font-lock keywords is required only when your buffer
was already set up and the change is interactive.  For all other cases,
it isn't only unnecessary, but even imposes a performance penalty, like
for C buffers.  This closes #3.